requirements.txt
Flask==2.3.3
PyYAML==6.0.1nemo-guardrails==0.11.0
python-dotenv==1.0.0
openai==1.0.0
requests==2.31.0